servicios

+++ DeberJson/Entidades/ProductoMsg.cs
@@ -13,7 +13,8 @@ namespace Entidades
         public int SupplierID { get; set; }
         public int CategoryID { get; set; }
         public string QuantityPerUnit { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int UnitsInStock { get; set; }
+     
     }
 }
